refactor skill rank card to be editable or not
