+++ apps/web/src/components/SkillRankCard/SkillRankCard.tsx
@@ -1,0 +1,86 @@
+import React from "react";
+import { useIntl } from "react-intl";
+
+import { UserSkill } from "@gc-digital-talent/graphql";
+import { Heading, HeadingLevel, Link, Separator } from "@gc-digital-talent/ui";
+
+import SkillRankListItem from "./SkillRankListItem";
+import NullMessage, { NullMessageProps } from "./NullMessage";
+
+interface SkillRankCardProps {
+  title: React.ReactNode;
+  description: React.ReactNode;
+  userSkills: UserSkill[];
+  titleAs?: HeadingLevel;
+  editable?: boolean;
+  editLink?: NullMessageProps["editLink"];
+}
+
+const SkillRankCard = ({
+  title,
+  description,
+  userSkills,
+  editLink,
+  editable = false,
+  titleAs = "h3",
+}: SkillRankCardProps) => {
+  const intl = useIntl();
+
+  return (
+    <div
+      data-h2-width="base(100%)"
+      data-h2-border-top="base(x.5 solid secondary)"
+      data-h2-shadow="base(xl)"
+      data-h2-radius="base(0 0 s s)"
+      data-h2-padding="base(x1)"
+      data-h2-background-color="base(foreground)"
+    >
+      <div
+        data-h2-display="base(flex)"
+        data-h2-align-items="base(center)"
+        data-h2-justify-content="base(space-between)"
+        data-h2-flex-direction="base(column) p-tablet(row)"
+        data-h2-margin-bottom="base(x1)"
+      >
+        <Heading level={titleAs} size="h6" data-h2-margin="base(0)">
+          {title}
+        </Heading>
+        {editable && editLink && (
+          <Link
+            mode="inline"
+            color="secondary"
+            href={editLink.href}
+            aria-label={editLink.label}
+          >
+            {intl.formatMessage({
+              defaultMessage: "Edit",
+              id: "igcM7/",
+              description: "Link text for editing a list of skill rankings",
+            })}
+          </Link>
+        )}
+      </div>
+      <p>{description}</p>
+      <Separator
+        orientation="horizontal"
+        data-h2-background-color="base(gray.lighter)"
+        data-h2-margin="base(x1 0)"
+      />
+      {userSkills.length ? (
+        <ul data-h2-margin="base(0)" data-h2-padding-left="base(x.75)">
+          {userSkills.map((userSkill) => (
+            <SkillRankListItem
+              key={userSkill.id}
+              userSkill={userSkill}
+              editable={editable}
+            />
+          ))}
+        </ul>
+      ) : (
+        <NullMessage editLink={editLink} editable={editable} />
+      )}
+    </div>
+  );
+};
+
+export default SkillRankCard;
